@@ -5,7 +5,7 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/matryer/moq/internal/registry"
+	"cirello.io/moq/internal/registry"
 )
 
 // Template is the Moq template. It is capable of generating the Moq
@@ -30,10 +30,9 @@ func (t Template) Execute(w io.Writer, data Data) error {
 	return t.tmpl.Execute(w, data)
 }
 
-// moqTemplate is the template for mocked code.
-// language=GoTemplate
+// moqTemplate is the template for mocked code (language=GoTemplate).
 var moqTemplate = `// Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package {{.PkgName}}
 
@@ -48,7 +47,7 @@ import (
 {{- if not $.SkipEnsure -}}
 // Ensure, that {{.MockName}} does implement {{$.SrcPkgQualifier}}{{.InterfaceName}}.
 // If this is not the case, regenerate this file with moq.
-var _ {{$.SrcPkgQualifier}}{{.InterfaceName -}} 
+var _ {{$.SrcPkgQualifier}}{{.InterfaceName -}}
 	{{- if .TypeParams }}[
 		{{- range $index, $param := .TypeParams}}
 			{{- if $index}}, {{end -}}
@@ -83,7 +82,7 @@ var _ {{$.SrcPkgQualifier}}{{.InterfaceName -}}
 //		// and then make assertions.
 //
 //	}
-type {{.MockName}} 
+type {{.MockName}}
 {{- if .TypeParams -}}
 	[{{- range $index, $param := .TypeParams}}
 			{{- if $index}}, {{end}}{{$param.Name | Exported}} {{$param.TypeString}}
@@ -116,7 +115,7 @@ func (mock *{{$mock.MockName}}
 	[{{- range $index, $param := $mock.TypeParams}}
 		{{- if $index}}, {{end}}{{$param.Name | Exported}}
 	{{- end -}}]
-{{- end -}}	
+{{- end -}}
 ) {{.Name}}({{.ArgList}}) {{.ReturnArgTypeList}} {
 {{- if not $.StubImpl}}
 	if mock.{{.Name}}Func == nil {
@@ -166,7 +165,7 @@ func (mock *{{$mock.MockName}}
 	[{{- range $index, $param := $mock.TypeParams}}
 		{{- if $index}}, {{end}}{{$param.Name | Exported}}
 	{{- end -}}]
-{{- end -}}	
+{{- end -}}
 ) {{.Name}}Calls() []struct {
 		{{- range .Params}}
 		{{.Name | Exported}} {{.TypeString}}
