@@ -1,5 +1,5 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package user
 
