@@ -1,8 +1,8 @@
 package importalias
 
 import (
-	srcclient "github.com/matryer/moq/pkg/moq/testpackages/importalias/source/client"
-	tgtclient "github.com/matryer/moq/pkg/moq/testpackages/importalias/target/client"
+	srcclient "cirello.io/moq/pkg/moq/testpackages/importalias/source/client"
+	tgtclient "cirello.io/moq/pkg/moq/testpackages/importalias/target/client"
 )
 
 type MiddleMan interface {
