@@ -1,11 +1,11 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package importalias
 
 import (
-	srcclient "github.com/matryer/moq/pkg/moq/testpackages/importalias/source/client"
-	tgtclient "github.com/matryer/moq/pkg/moq/testpackages/importalias/target/client"
+	srcclient "cirello.io/moq/pkg/moq/testpackages/importalias/source/client"
+	tgtclient "cirello.io/moq/pkg/moq/testpackages/importalias/target/client"
 	"sync"
 )
 
