@@ -1,10 +1,10 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package two
 
 import (
-	"github.com/matryer/moq/pkg/moq/testpackages/imports/one"
+	"cirello.io/moq/pkg/moq/testpackages/imports/one"
 	"sync"
 )
 
