@@ -1,12 +1,12 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package two
 
 import (
 	"sync"
 
-	"github.com/matryer/moq/pkg/moq/testpackages/imports/one"
+	"cirello.io/moq/pkg/moq/testpackages/imports/one"
 )
 
 // Ensure, that goimportsMock does implement DoSomething.
