@@ -1,10 +1,10 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package mock
 
 import (
-	shadowhttp "github.com/matryer/moq/pkg/moq/testpackages/shadow/http"
+	shadowhttp "cirello.io/moq/pkg/moq/testpackages/shadow/http"
 	nethttp "net/http"
 	"sync"
 )
