@@ -1,10 +1,10 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package shadowtypes
 
 import (
-	"github.com/matryer/moq/pkg/moq/testpackages/shadowtypes/types"
+	"cirello.io/moq/pkg/moq/testpackages/shadowtypes/types"
 	"sync"
 )
 
