@@ -1,10 +1,10 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package syncimport
 
 import (
-	"github.com/matryer/moq/pkg/moq/testpackages/syncimport/sync"
+	"cirello.io/moq/pkg/moq/testpackages/syncimport/sync"
 	stdsync "sync"
 )
 
