@@ -1,11 +1,11 @@
 package base
 
 import (
-	four "github.com/matryer/moq/pkg/moq/testpackages/transientimport/four/app/v1"
-	one "github.com/matryer/moq/pkg/moq/testpackages/transientimport/one/v1"
-	"github.com/matryer/moq/pkg/moq/testpackages/transientimport/onev1"
-	three "github.com/matryer/moq/pkg/moq/testpackages/transientimport/three/v1"
-	two "github.com/matryer/moq/pkg/moq/testpackages/transientimport/two/app/v1"
+	four "cirello.io/moq/pkg/moq/testpackages/transientimport/four/app/v1"
+	one "cirello.io/moq/pkg/moq/testpackages/transientimport/one/v1"
+	"cirello.io/moq/pkg/moq/testpackages/transientimport/onev1"
+	three "cirello.io/moq/pkg/moq/testpackages/transientimport/three/v1"
+	two "cirello.io/moq/pkg/moq/testpackages/transientimport/two/app/v1"
 )
 
 type Transient interface {
