@@ -1,7 +1,7 @@
 package transientimport
 
 import (
-	"github.com/matryer/moq/pkg/moq/testpackages/transientimport/base"
+	"cirello.io/moq/pkg/moq/testpackages/transientimport/base"
 )
 
 type Transient = base.Transient
