@@ -1,14 +1,14 @@
 // Code generated by moq; DO NOT EDIT.
-// github.com/matryer/moq
+// cirello.io/moq
 
 package transientimport
 
 import (
-	fourappv1 "github.com/matryer/moq/pkg/moq/testpackages/transientimport/four/app/v1"
-	transientimportonev1 "github.com/matryer/moq/pkg/moq/testpackages/transientimport/one/v1"
-	testpackagestransientimportonev1 "github.com/matryer/moq/pkg/moq/testpackages/transientimport/onev1"
-	threev1 "github.com/matryer/moq/pkg/moq/testpackages/transientimport/three/v1"
-	twoappv1 "github.com/matryer/moq/pkg/moq/testpackages/transientimport/two/app/v1"
+	fourappv1 "cirello.io/moq/pkg/moq/testpackages/transientimport/four/app/v1"
+	transientimportonev1 "cirello.io/moq/pkg/moq/testpackages/transientimport/one/v1"
+	testpackagestransientimportonev1 "cirello.io/moq/pkg/moq/testpackages/transientimport/onev1"
+	threev1 "cirello.io/moq/pkg/moq/testpackages/transientimport/three/v1"
+	twoappv1 "cirello.io/moq/pkg/moq/testpackages/transientimport/two/app/v1"
 	"sync"
 )
 
